+++ src/main/java/com/jh/hotelbookingmanagement/repository/RoomStatusRepository.java
@@ -3,5 +3,5 @@ package com.jh.hotelbookingmanagement.repository;
 import com.jh.hotelbookingmanagement.entity.RoomStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomStatusRepository extends JpaRepository<RoomStatus, Integer> {
+public interface RoomStatusRepository extends JpaRepository<RoomStatus, Long> {
 }
